Use schema versions for queries. * Use actual schema for query in TX. * Pass schema version to remote fragment. * Wait for catalog schema ready on the remote side. * Planner cache use catalog version.

+++ modules/runner/src/integrationTest/java/org/apache/ignite/internal/sql/api/ItCommonApiTest.java
@@ -242,7 +242,7 @@ public class ItCommonApiTest extends ClusterPerClassIntegrationTest {
 
         /** {@inheritDoc} */
         @Override
-        public CompletableFuture<SchemaPlus> actualSchemaAsync(long ver) {
+        public CompletableFuture<SchemaPlus> actualSchemaAsync(int ver) {
             throw new UnsupportedOperationException();
         }
